Added post requests and IDs on devices

+++ api/functions/config_files_manager.js
@@ -1,4 +1,5 @@
 const fs = require('fs').promises;
+const { v4: uuidv4 } = require('uuid');
 
 async function getUrlById(filePath, id) {
     try {
@@ -16,34 +17,39 @@ async function getUrlById(filePath, id) {
     }
 }
 
-async function addUrl(filePath, id, url) {
-    try {
-        const data = await fs.readFile(filePath, 'utf8');
-        const jsonObject = JSON.parse(data);
+async function addUrl(filePath, description, url) {
+    const data = await fs.readFile(filePath, 'utf8');
+    const jsonObject = JSON.parse(data);
 
-        jsonObject.push({ id: id, url: url });
-
-        await fs.writeFile(filePath, JSON.stringify(jsonObject));
-    } catch (err) {
-        console.error('Error:', err);
+    const urlExists = jsonObject.some(item => item.url === url);
+    if (urlExists) {
+        throw new Error('URL already exists');
     }
+
+    const id = uuidv4();
+
+    jsonObject.push({ id, description, url });
+
+    await fs.writeFile(filePath, JSON.stringify(jsonObject));
+
+    return id;
 }
 
-async function removeUrl(id) {
+async function removeUrl(filePath, id) {
     try {
         const data = await fs.readFile(filePath, 'utf8');
         const jsonObject = JSON.parse(data);
 
-        for (let i = 0; i < jsonObject.length; i++) {
-            if (jsonObject[i].id === id) {
-                jsonObject.splice(i, 1);
-                break;
-            }
+        const index = jsonObject.findIndex(item => item.id === id);
+        if (index !== -1) {
+            jsonObject.splice(index, 1);
+            await fs.writeFile(filePath, JSON.stringify(jsonObject));
+        } else {
+            throw new Error('ID not found');
         }
-
-        await fs.writeFile(filePath, JSON.stringify(jsonObject));
     } catch (err) {
         console.error('Error:', err);
+        throw err;
     }
 }
